Move localHelpers to its own folder
